fixed up some css and updated klicky.ts
