+++ index.html
@@ -149,7 +149,7 @@
                 <p>Events</p>
               </div>
               <div class="add">
-                <img src="./assets/20340289291547546467-128.png" width="15">
+                <img src="./assets/20340289291547546467-128.png" width="20">
               </div>
             </div>
           </div>
